[Raj] feat. developing crud service methods with mysql database

+++ EmployeePayrollApp/src/main/java/com/bridgelabz/employeepayrollapp/controller/EmployeePayrollController.java
@@ -22,12 +22,7 @@ public class EmployeePayrollController {
     @Autowired
     private EmployeePayrollService employeePayrollService;  
 
-//    @PostMapping("/create")
-//    public ResponseEntity<?> createEmployee(@Valid @RequestBody EmployeeDTO employeeDTO) {
-//        System.out.println("Received Employee DTO: " + employeeDTO);
-//        Employee employee = employeePayrollService.createEmployee(employeeDTO);
-//        return new ResponseEntity<>(employee, HttpStatus.CREATED);
-//    }
+
     @PostMapping("/create")
     public ResponseEntity<?> createEmployee(@Valid @RequestBody EmployeeDTO employeeDTO) {
         System.out.println("Received Employee DTO: " + employeeDTO); // Debugging line
@@ -35,26 +30,23 @@ public class EmployeePayrollController {
         return new ResponseEntity<>(employee, HttpStatus.CREATED);
     }
 
-
-
-
     @GetMapping("/")
     public List<Employee> getAllEmployees() {
         return employeePayrollService.getAllEmployees();
     }
 
     @GetMapping("/get/{id}")
-    public Employee getEmployeeById(@PathVariable("id") Long id) {
+    public Employee getEmployeeById(@PathVariable("id") int id) {
         return employeePayrollService.getEmployeeById(id);
     }
     
     @PutMapping("/update/{id}")
-    public Employee updateEmployee(@PathVariable("id") Long id,@Valid @RequestBody EmployeeDTO employeeDTO) {
+    public Employee updateEmployee(@PathVariable("id") int id,@Valid @RequestBody EmployeeDTO employeeDTO) {
         return employeePayrollService.updateEmployee(id, employeeDTO);
     }
     
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<ResponseDTO> deleteEmployee(@PathVariable("id") Long id) {
+    public ResponseEntity<ResponseDTO> deleteEmployee(@PathVariable("id") int id) {
         employeePayrollService.deleteEmployee(id);
         ResponseDTO response = new ResponseDTO("Employee deleted successfully", "Employee ID: " + id);
         return new ResponseEntity<>(response, HttpStatus.OK);
